Add Toggle & Remove Buttons, Handle Duplicates
In this commit, I made improvements to the displayBook() function to
display books in the library more efficiently. I replaced the previous
for loop with a more concise one, and now each book is appended to the
DOM as it is iterated through, reducing unnecessary repetitions.

Additionally, I added the functionality for the "Read/Not Read" toggle
button and the "Remove" button in the displayBook() function.